[Fix] some refactoring in auth service
